+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [#1645](https://github.com/hyperf/hyperf/pull/1645) Added parameter injection support for closure route.
 - [#1647](https://github.com/hyperf/hyperf/pull/1647) Added RedisStringHandler for model-cache.
+- [#1654](https://github.com/hyperf/hyperf/pull/1654) Added `RenderException` to catch render exceptions in view.
 
 ## Fixed